+++ crates/model/src/events/order/filled.rs
@@ -333,6 +333,10 @@ impl OrderEvent for OrderFilled {
         Some(self.last_qty)
     }
 
+    fn activation_price(&self) -> Option<Price> {
+        None
+    }
+
     fn trigger_price(&self) -> Option<Price> {
         None
     }